+++ connect-example/src/main/java/com/stripe/android/connect/example/data/EmbeddedComponentManagerFactory.kt
@@ -30,7 +30,7 @@ class EmbeddedComponentManagerFactory @Inject constructor(
 
         return EmbeddedComponentManager(
             publishableKey = publishableKey,
-            clientSecretProvider = ::fetchClientSecret,
+            fetchClientSecret = ::fetchClientSecret,
             customFonts = listOf(
                 CustomFontSource(
                     assetsFilePath = "fonts/doto.ttf",