+++ paymentsheet/src/androidTest/java/com/stripe/android/paymentelement/EmbeddedPaymentElementAnalyticsTest.kt
@@ -32,14 +32,18 @@ import org.junit.Rule
 import org.junit.Test
 import kotlin.time.Duration.Companion.seconds
 
+@OptIn(ExperimentalAnalyticEventCallbackApi::class)
 internal class EmbeddedPaymentElementAnalyticsTest {
     private val networkRule = NetworkRule(
         hostsToTrack = listOf(ApiRequest.API_HOST, AnalyticsRequest.HOST),
         validationTimeout = 1.seconds, // Analytics requests happen async.
     )
+    private val analyticEventRule = AnalyticEventRule()
 
     @get:Rule
-    val testRules: TestRules = TestRules.create(networkRule = networkRule)
+    val testRules: TestRules = TestRules.create(networkRule = networkRule) {
+        around(analyticEventRule)
+    }
 
     private val embeddedContentPage = EmbeddedContentPage(testRules.compose)
     private val formPage = EmbeddedFormPage(testRules.compose)
@@ -69,6 +73,7 @@ internal class EmbeddedPaymentElementAnalyticsTest {
             CreateIntentResult.Success("pi_example_secret_12345")
         },
         resultCallback = ::assertCompleted,
+        analyticEventCallback = analyticEventRule,
     ) { testContext ->
         networkRule.enqueue(
             host("api.stripe.com"),
@@ -91,6 +96,9 @@ internal class EmbeddedPaymentElementAnalyticsTest {
         validateAnalyticsRequest(eventName = "mc_card_number_completed")
 
         testContext.configure()
+
+        analyticEventRule.assertMatchesExpectedEvent(AnalyticEvent.PresentedSheet())
+
         embeddedContentPage.clickOnLpm("card")
         formPage.fillOutCardDetails()
 
@@ -143,6 +151,7 @@ internal class EmbeddedPaymentElementAnalyticsTest {
             CreateIntentResult.Success("pi_example_secret_12345")
         },
         resultCallback = ::assertCompleted,
+        analyticEventCallback = analyticEventRule,
     ) { testContext ->
         networkRule.enqueue(
             host("api.stripe.com"),
@@ -163,6 +172,9 @@ internal class EmbeddedPaymentElementAnalyticsTest {
         testContext.configure {
             customer(PaymentSheet.CustomerConfiguration("cus_123", "ek_test"))
         }
+
+        analyticEventRule.assertMatchesExpectedEvent(AnalyticEvent.PresentedSheet())
+
         embeddedContentPage.assertHasSelectedSavedPaymentMethod("pm_12345")
 
         networkRule.enqueue(
@@ -204,6 +216,7 @@ internal class EmbeddedPaymentElementAnalyticsTest {
             CreateIntentResult.Success("pi_example_secret_12345")
         },
         resultCallback = ::assertCompleted,
+        analyticEventCallback = analyticEventRule,
     ) { testContext ->
         networkRule.enqueue(
             host("api.stripe.com"),
@@ -224,6 +237,8 @@ internal class EmbeddedPaymentElementAnalyticsTest {
         testContext.configure {
             customer(PaymentSheet.CustomerConfiguration("cus_123", "ek_test"))
         }
+
+        analyticEventRule.assertMatchesExpectedEvent(AnalyticEvent.PresentedSheet())
 
         validateAnalyticsRequest(eventName = "mc_embedded_manage_savedpm_show")
         embeddedContentPage.clickViewMore()