+++ paymentsheet/src/androidTest/java/com/stripe/android/paymentsheet/utils/FlowControllerTestRunner.kt
@@ -8,6 +8,8 @@ import com.google.common.truth.Truth.assertThat
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.link.account.LinkStore
 import com.stripe.android.networktesting.NetworkRule
+import com.stripe.android.paymentelement.AnalyticEventCallback
+import com.stripe.android.paymentelement.ExperimentalAnalyticEventCallbackApi
 import com.stripe.android.paymentsheet.CreateIntentCallback
 import com.stripe.android.paymentsheet.MainActivity
 import com.stripe.android.paymentsheet.PaymentOptionsActivity
@@ -54,12 +56,35 @@ internal fun runFlowControllerTest(
     resultCallback: PaymentSheetResultCallback,
     block: suspend (FlowControllerTestRunnerContext) -> Unit,
 ) {
+    @OptIn(ExperimentalAnalyticEventCallbackApi::class)
+    runFlowControllerTest(
+        networkRule = networkRule,
+        integrationType = integrationType,
+        callConfirmOnPaymentOptionCallback = callConfirmOnPaymentOptionCallback,
+        createIntentCallback = createIntentCallback,
+        analyticEventCallback = null,
+        resultCallback = resultCallback,
+        block = block
+    )
+}
+
+@OptIn(ExperimentalAnalyticEventCallbackApi::class)
+internal fun runFlowControllerTest(
+    networkRule: NetworkRule,
+    integrationType: IntegrationType = IntegrationType.Compose,
+    callConfirmOnPaymentOptionCallback: Boolean = true,
+    createIntentCallback: CreateIntentCallback? = null,
+    analyticEventCallback: AnalyticEventCallback? = null,
+    resultCallback: PaymentSheetResultCallback,
+    block: suspend (FlowControllerTestRunnerContext) -> Unit,
+) {
     val countDownLatch = CountDownLatch(1)
     val configureCallbackTurbine = Turbine<PaymentOption?>()
 
     val factory = FlowControllerTestFactory(
         callConfirmOnPaymentOptionCallback = callConfirmOnPaymentOptionCallback,
         createIntentCallback = createIntentCallback,
+        analyticEventCallback = analyticEventCallback,
         configureCallbackTurbine = configureCallbackTurbine,
         resultCallback = { result ->
             resultCallback.onPaymentSheetResult(result)