+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/UpdatePaymentMethodInteractor.kt
@@ -36,13 +36,13 @@ internal interface UpdatePaymentMethodInteractor {
     val setAsDefaultCheckboxEnabled: Boolean
     val shouldShowSaveButton: Boolean
     val allowCardEdit: Boolean
+    val editCardDetailsInteractor: EditCardDetailsInteractor
 
     val state: StateFlow<State>
 
     data class State(
         val error: ResolvableString?,
         val status: Status,
-        val cardBrandChoice: CardBrandChoice,
         val setAsDefaultCheckboxChecked: Boolean,
         val isSaveButtonEnabled: Boolean,
     )
@@ -57,7 +57,6 @@ internal interface UpdatePaymentMethodInteractor {
 
     sealed class ViewAction {
         data object RemovePaymentMethod : ViewAction()
-        data class BrandChoiceChanged(val cardBrandChoice: CardBrandChoice) : ViewAction()
         data object SaveButtonPressed : ViewAction()
         data class SetAsDefaultCheckboxChanged(val isChecked: Boolean) : ViewAction()
     }
@@ -81,6 +80,7 @@ internal typealias UpdateCardPaymentMethodOperation = suspend (
     paymentMethod: PaymentMethod,
     cardUpdateParams: CardUpdateParams
 ) -> Result<PaymentMethod>
+
 internal typealias PaymentMethodSetAsDefaultOperation = suspend (
     paymentMethod: PaymentMethod
 ) -> Result<Unit>
@@ -97,16 +97,14 @@ internal class DefaultUpdatePaymentMethodInteractor(
     private val setDefaultPaymentMethodExecutor: PaymentMethodSetAsDefaultOperation,
     private val onBrandChoiceSelected: (CardBrand) -> Unit,
     private val onUpdateSuccess: () -> Unit,
-    workContext: CoroutineContext = Dispatchers.Default,
+    private val workContext: CoroutineContext = Dispatchers.Default,
 ) : UpdatePaymentMethodInteractor {
     private val coroutineScope = CoroutineScope(workContext + SupervisorJob())
     private val error = MutableStateFlow(getInitialError())
     private val status = MutableStateFlow(UpdatePaymentMethodInteractor.Status.Idle)
-    private val cardBrandChoice = MutableStateFlow(getInitialCardBrandChoice())
-    private val cardBrandHasBeenChanged = MutableStateFlow(false)
     private val initialSetAsDefaultCheckedValue = isDefaultPaymentMethod
     private val setAsDefaultCheckboxChecked = MutableStateFlow(initialSetAsDefaultCheckedValue)
-    private val savedCardBrand = MutableStateFlow(getInitialCardBrandChoice())
+    private val cardUpdateParams = MutableStateFlow<CardUpdateParams?>(null)
 
     // We don't yet support setting SEPA payment methods as defaults, so we hide the checkbox for now.
     override val shouldShowSetAsDefaultCheckbox = (
@@ -135,27 +133,42 @@ internal class DefaultUpdatePaymentMethodInteractor(
     private val _setAsDefaultValueChanged = setAsDefaultCheckboxChecked.mapAsStateFlow { setAsDefaultCheckboxChecked ->
         setAsDefaultCheckboxChecked != initialSetAsDefaultCheckedValue
     }
+    override val editCardDetailsInteractor by lazy {
+        if (displayableSavedPaymentMethod.savedPaymentMethod is SavedPaymentMethod.Card) {
+            EditCardDetailsInteractor.create(
+                card = displayableSavedPaymentMethod.savedPaymentMethod.card,
+                onCardUpdateParamsChanged = ::onCardUpdateParamsChanged,
+                coroutineScope = coroutineScope,
+                isModifiable = displayableSavedPaymentMethod.isModifiable(),
+                cardBrandFilter = cardBrandFilter,
+                onBrandChoiceChanged = onBrandChoiceSelected,
+            )
+        } else {
+            throw IllegalStateException("card payment method required for creating editCardDetailsInteractor")
+        }
+    }
+
+    private fun onCardUpdateParamsChanged(cardUpdateParams: CardUpdateParams?) {
+        this.cardUpdateParams.value = cardUpdateParams
+    }
 
     private val _state = combineAsStateFlow(
         error,
         status,
-        cardBrandChoice,
-        cardBrandHasBeenChanged,
         setAsDefaultCheckboxChecked,
         _setAsDefaultValueChanged,
+        cardUpdateParams
     ) { error,
         status,
-        cardBrandChoice,
-        cardBrandHasBeenChanged,
         setAsDefaultCheckboxChecked,
-        setAsDefaultValueChanged ->
-        val isSaveButtonEnabled = (cardBrandHasBeenChanged || setAsDefaultValueChanged) &&
+        setAsDefaultValueChanged,
+        cardUpdateParams ->
+        val isSaveButtonEnabled = (setAsDefaultValueChanged || cardUpdateParams != null) &&
             status == UpdatePaymentMethodInteractor.Status.Idle
 
         UpdatePaymentMethodInteractor.State(
             error = error,
             status = status,
-            cardBrandChoice = cardBrandChoice,
             isSaveButtonEnabled = isSaveButtonEnabled,
             setAsDefaultCheckboxChecked = isDefaultPaymentMethod || setAsDefaultCheckboxChecked,
         )
@@ -166,9 +179,6 @@ internal class DefaultUpdatePaymentMethodInteractor(
         when (viewAction) {
             UpdatePaymentMethodInteractor.ViewAction.RemovePaymentMethod -> removePaymentMethod()
             UpdatePaymentMethodInteractor.ViewAction.SaveButtonPressed -> savePaymentMethod()
-            is UpdatePaymentMethodInteractor.ViewAction.BrandChoiceChanged -> onBrandChoiceChanged(
-                viewAction.cardBrandChoice
-            )
             is UpdatePaymentMethodInteractor.ViewAction.SetAsDefaultCheckboxChanged -> onSetAsDefaultCheckboxChanged(
                 isChecked = viewAction.isChecked
             )
@@ -192,7 +202,7 @@ internal class DefaultUpdatePaymentMethodInteractor(
             error.emit(getInitialError())
             status.emit(UpdatePaymentMethodInteractor.Status.Updating)
 
-            val updateCardBrandResult = maybeUpdateCardBrand()
+            val updateCardBrandResult = maybeUpdateCard()
             val setDefaultPaymentMethodResult = maybeSetDefaultPaymentMethod()
 
             val updateResult = getUpdateResult(
@@ -210,17 +220,14 @@ internal class DefaultUpdatePaymentMethodInteractor(
         }
     }
 
-    private suspend fun maybeUpdateCardBrand(): Result<PaymentMethod>? {
-        val newCardBrand = cardBrandChoice.value.brand
-        return if (cardBrandHasBeenChanged.value) {
+    private suspend fun maybeUpdateCard(): Result<PaymentMethod>? {
+        val cardUpdateParams = cardUpdateParams.value
+        return if (cardUpdateParams != null) {
             updatePaymentMethodExecutor(
                 displayableSavedPaymentMethod.paymentMethod,
-                CardUpdateParams(
-                    cardBrand = newCardBrand
-                )
+                cardUpdateParams
             ).onSuccess {
-                savedCardBrand.emit(CardBrandChoice(brand = newCardBrand, enabled = true))
-                cardBrandHasBeenChanged.emit(false)
+                this.cardUpdateParams.value = null
             }
         } else {
             null
@@ -254,25 +261,8 @@ internal class DefaultUpdatePaymentMethodInteractor(
         }
     }
 
-    private fun onBrandChoiceChanged(cardBrandChoice: CardBrandChoice) {
-        this.cardBrandChoice.value = cardBrandChoice
-        val changed = cardBrandChoice != savedCardBrand.value
-        this.cardBrandHasBeenChanged.value = changed
-
-        if (changed) {
-            onBrandChoiceSelected(cardBrandChoice.brand)
-        }
-    }
-
     private fun onSetAsDefaultCheckboxChanged(isChecked: Boolean) {
         setAsDefaultCheckboxChecked.update { isChecked }
-    }
-
-    private fun getInitialCardBrandChoice(): CardBrandChoice {
-        return when (val savedPaymentMethod = displayableSavedPaymentMethod.savedPaymentMethod) {
-            is SavedPaymentMethod.Card -> savedPaymentMethod.card.getPreferredChoice(cardBrandFilter)
-            else -> CardBrandChoice(brand = CardBrand.Unknown, enabled = true)
-        }
     }
 
     private fun paymentMethodIsExpiredCard(): Boolean {