+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentSheetViewModelTest.kt
@@ -118,6 +118,8 @@ import com.stripe.android.paymentsheet.state.WalletsProcessingState
 import com.stripe.android.paymentsheet.ui.CardBrandChoice
 import com.stripe.android.paymentsheet.ui.PrimaryButton
 import com.stripe.android.paymentsheet.ui.UpdatePaymentMethodInteractor
+import com.stripe.android.paymentsheet.ui.editCardDetailsInteractorHelper
+import com.stripe.android.paymentsheet.ui.updateCardBrand
 import com.stripe.android.paymentsheet.utils.LinkTestUtils
 import com.stripe.android.paymentsheet.utils.prefillCreate
 import com.stripe.android.paymentsheet.utils.prefilledBuilder
@@ -154,7 +156,6 @@ import org.junit.runner.RunWith
 import org.mockito.MockitoAnnotations
 import org.mockito.kotlin.any
 import org.mockito.kotlin.anyOrNull
-import org.mockito.kotlin.argThat
 import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.doReturn
@@ -287,14 +288,9 @@ internal class PaymentSheetViewModelTest {
             if (currentScreen is PaymentSheetScreen.UpdatePaymentMethod) {
                 val interactor = currentScreen.interactor
 
-                interactor.handleViewAction(
-                    UpdatePaymentMethodInteractor.ViewAction.BrandChoiceChanged(
-                        CardBrandChoice(
-                            CardBrand.Visa,
-                            enabled = true
-                        )
-                    )
-                )
+                interactor.editCardDetailsInteractorHelper {
+                    updateCardBrand(CardBrand.Visa)
+                }
 
                 interactor.handleViewAction(UpdatePaymentMethodInteractor.ViewAction.SaveButtonPressed)
             }
@@ -323,6 +319,7 @@ internal class PaymentSheetViewModelTest {
     @Suppress("LongMethod")
     fun `modifyPaymentMethod updates payment methods and sends event on successful update`() = runTest {
         Dispatchers.setMain(testDispatcher)
+        val eventReporter = FakeEventReporter()
         val paymentMethods = PaymentMethodFixtures.createCards(5)
 
         val firstPaymentMethod = paymentMethods.first()
@@ -344,7 +341,8 @@ internal class PaymentSheetViewModelTest {
         )
         val viewModel = createViewModel(
             customer = EMPTY_CUSTOMER_STATE.copy(paymentMethods = paymentMethods),
-            customerRepository = customerRepository
+            customerRepository = customerRepository,
+            eventReporter = eventReporter
         )
 
         viewModel.navigationHandler.currentScreen.test {
@@ -361,14 +359,9 @@ internal class PaymentSheetViewModelTest {
             if (currentScreen is PaymentSheetScreen.UpdatePaymentMethod) {
                 val interactor = currentScreen.interactor
 
-                interactor.handleViewAction(
-                    UpdatePaymentMethodInteractor.ViewAction.BrandChoiceChanged(
-                        CardBrandChoice(
-                            CardBrand.Visa,
-                            enabled = true
-                        )
-                    )
-                )
+                interactor.editCardDetailsInteractorHelper {
+                    updateCardBrand(CardBrand.Visa)
+                }
 
                 interactor.handleViewAction(UpdatePaymentMethodInteractor.ViewAction.SaveButtonPressed)
             }
@@ -376,7 +369,9 @@ internal class PaymentSheetViewModelTest {
             assertThat(awaitItem()).isInstanceOf<SelectSavedPaymentMethods>()
         }
 
-        verify(eventReporter).onUpdatePaymentMethodSucceeded(CardBrand.Visa)
+        val updatePaymentMethodSucceededCall = eventReporter.updatePaymentMethodSucceededCalls.awaitItem()
+        eventReporter.updatePaymentMethodSucceededCalls.ensureAllEventsConsumed()
+        assertThat(updatePaymentMethodSucceededCall.selectedBrand).isEqualTo(CardBrand.Visa)
 
         val idCaptor = argumentCaptor<String>()
         val paramsCaptor = argumentCaptor<PaymentMethodUpdateParams>()
@@ -406,6 +401,8 @@ internal class PaymentSheetViewModelTest {
 
     @Test
     fun `modifyPaymentMethod sends event on failed update`() = runTest {
+        Dispatchers.setMain(testDispatcher)
+        val eventReporter = FakeEventReporter()
         val paymentMethods = PaymentMethodFixtures.createCards(5)
 
         val firstPaymentMethod = paymentMethods.first()
@@ -419,7 +416,8 @@ internal class PaymentSheetViewModelTest {
         )
         val viewModel = createViewModel(
             customer = EMPTY_CUSTOMER_STATE.copy(paymentMethods = paymentMethods),
-            customerRepository = customerRepository
+            customerRepository = customerRepository,
+            eventReporter = eventReporter
         )
 
         viewModel.navigationHandler.currentScreen.test {
@@ -436,25 +434,18 @@ internal class PaymentSheetViewModelTest {
             if (currentScreen is PaymentSheetScreen.UpdatePaymentMethod) {
                 val interactor = currentScreen.interactor
 
-                interactor.handleViewAction(
-                    UpdatePaymentMethodInteractor.ViewAction.BrandChoiceChanged(
-                        CardBrandChoice(
-                            CardBrand.Visa,
-                            enabled = true
-                        )
-                    )
-                )
+                interactor.editCardDetailsInteractorHelper {
+                    updateCardBrand(CardBrand.Visa)
+                }
 
                 interactor.handleViewAction(UpdatePaymentMethodInteractor.ViewAction.SaveButtonPressed)
             }
         }
 
-        verify(eventReporter).onUpdatePaymentMethodFailed(
-            selectedBrand = eq(CardBrand.Visa),
-            error = argThat {
-                message == "No network found!"
-            }
-        )
+        val onUpdatePaymentMethodFailedCall = eventReporter.updatePaymentMethodFailedCalls.awaitItem()
+        eventReporter.updatePaymentMethodFailedCalls.ensureAllEventsConsumed()
+        assertThat(onUpdatePaymentMethodFailedCall.selectedBrand).isEqualTo(CardBrand.Visa)
+        assertThat(onUpdatePaymentMethodFailedCall.error.message).isEqualTo("No network found!")
     }
 
     @Test
@@ -2998,15 +2989,9 @@ internal class PaymentSheetViewModelTest {
 
             if (currentScreen is PaymentSheetScreen.UpdatePaymentMethod) {
                 val interactor = currentScreen.interactor
-
-                interactor.handleViewAction(
-                    UpdatePaymentMethodInteractor.ViewAction.BrandChoiceChanged(
-                        CardBrandChoice(
-                            CardBrand.Visa,
-                            enabled = true
-                        )
-                    )
-                )
+                interactor.editCardDetailsInteractorHelper {
+                    updateCardBrand(CardBrand.Visa)
+                }
 
                 verify(customerRepository, never()).updatePaymentMethod(any(), any(), any())
             }